Clown Car Interior Scene WITH SOUND ISSUES
I can't for the life of me get the clowns horn to sound at the proper
time when he uses it to "respond" to the player. Spent all night trying
to get it to work.. YA'D THINK THIS WOULD BE AN EASY THING.. But.. guess
not? Wouldnt work in an actionlist.. the sound would play at the end of
ALL the dialog.. ONCE.. Tried to attach sound component to the clown, &
a script so i could add an event to the horn honking animation, but
couldnt get it to work. I NEED YOUR BRAINS! HELP!